src/requirements.txt
appdirs==1.4.4
attrs==20.2.0
Automat==20.2.0
black==20.8b1
cffi==1.14.3
click==7.1.2
constantly==15.1.0
cryptography==3.1.1
cssselect==1.1.0
decorator==4.4.2
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.0.0
incremental==17.5.0
iniconfig==1.0.1
itemadapter==0.1.1
itemloaders==1.0.3
jmespath==0.10.0
lxml==4.5.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.21
packaging==20.4
parsel==1.6.0
pathspec==0.8.0
pluggy==0.13.1
Protego==0.1.16
protobuf==3.13.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.1.1
queuelib==1.5.0
regex==2020.9.27
Scrapy==2.3.0
service-identity==18.1.0
six==1.15.0
toml==0.10.1
Twisted==20.3.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
validators==0.18.1
w3lib==1.22.0
zipp==3.3.0
zope.interface==5.1.2
